--- src/Component/Layout/Footer/Footer.css
@@ -1,7 +1,0 @@
-#sticky-footer {
-    flex-shrink: none;
-  }
-
-.footer{
-  bottom: 0;
-}
