--- src/Component/Layout/Footer/Footer.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import './Footer.css';
-
-function Footer() {
-  return (
-    <footer id="sticky-footer" className="footer py-4 bg-light text-white-50 ">
-      <div className="container text-center text-secondary">
-        <small>
-          Universidad Nacional de Quilmes, Ciencia y Tecnologia, Tecnicatura Universitaria en Programación Informática.
-        </small>
-      </div>
-    </footer>
-  );
-}
-
-export default Footer;
