--- src/Component/Layout/Nav/Nav.css
@@ -1,4 +1,0 @@
-.nav-logo {
-    height: 10vmin;
-    pointer-events: none;
-  }
