--- src/Component/Layout/Sidebar/Sidebar.css
@@ -1,34 +1,0 @@
-li.active{
-    background-color: #428bca;
-}
-li {
-    padding-bottom: 5px;
-    padding-top: 5px;
-}
-
-.sidebar{
-    background-color: #f5f5f5;
-    padding-right: 20px;
-    padding-top: 20px;
-    min-height: 100%;
-}
-
-/* Sidebar navigation */
-.nav-sidebar {
-    margin-right: -21px; /* 20px padding + 1px border */
-    margin-bottom: 20px;
-    margin-left: -20px;
-}
-.nav-sidebar > li > a {
-    padding-right: 20px;
-    padding-left: 20px;
-}
-.nav-sidebar > .active > a,
-.nav-sidebar > .active > a:hover,
-.nav-sidebar > .active > a:focus {
-    color: #fff;
-    background-color: #428bca;
-}
-button{
-    color: #fafafa
-}
