+++ src/Component/dashboard/Dashboard.css
@@ -1,0 +1,6 @@
+.footer{
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 2.5rem; 
+  }
