+++ src/Component/dashboard/Nav.jsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import clsx from 'clsx';
+
+import {
+  IconButton, AppBar, Toolbar, Typography, Badge,
+} from '@material-ui/core';
+import MenuIcon from '@material-ui/icons/Menu';
+import NotificationsIcon from '@material-ui/icons/Notifications';
+
+
+export default function Nav({ classes, handleDrawerOpen, open }) {
+  return (
+    <div>
+      <AppBar position="absolute" className={clsx(classes.appBar, open && classes.appBarShift)}>
+        <Toolbar className={classes.toolbar}>
+          <IconButton
+            edge="start"
+            color="inherit"
+            aria-label="open drawer"
+            onClick={handleDrawerOpen}
+            className={clsx(classes.menuButton, open && classes.menuButtonHidden)}
+          >
+            <MenuIcon />
+          </IconButton>
+          <Typography component="h1" variant="h6" color="inherit" noWrap className={classes.title}>
+                UNQuivalencias
+          </Typography>
+          <IconButton color="inherit">
+            <Badge badgeContent={2} color="secondary">
+              <NotificationsIcon />
+            </Badge>
+          </IconButton>
+        </Toolbar>
+      </AppBar>
+    </div>
+  );
+}
